requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
psycopg2==2.9.3
pygame==2.1.2
python-magic==0.4.24
pytz==2021.3
requests==2.27.1
RPi.GPIO==0.7.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
